final commit

+++ client/src/App.js
@@ -6,12 +6,14 @@ import UpdateTodo from './components/updateTodo';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faTag } from '@fortawesome/free-solid-svg-icons'
 import splitByComma from './utility/splitByComma';
+import Modal from 'react-modal'
 
 function App() {
 	const [todolist, setTodolist] = useState([])
 	const [update, setUpdate] = useState(false)
 	const [id, setId] = useState("");
 	const [open, setOpen] = useState(false);
+	const [openModal, setOpenModal] = useState(false);
 	const [completed, setCompleted] = useState(false)
 	const [editData, setEditData] = useState({ title: '', category: "" })
 
@@ -28,6 +30,11 @@ function App() {
 		})
 	}
 
+	function handleClick(e) {
+		setOpenModal(false)
+		console.log(e.target.name)
+	}
+
 	function handleEdit(e) {
 		setEditData(splitByComma(e.target.value))
 		setId(e.target.name);
@@ -40,20 +47,21 @@ function App() {
 	}
 
 	function handleDelete(e) {
-		alert('Deleting ' + e.target.value)
-		axios.delete(`http://localhost:8000/api/todo/${e.target.name}`,)
-		setTodolist((data) => {
-			return data.filter((todo) => todo._id !== e.target.name)
-		})
+		if (window.confirm("Deleting task ", e.target.name)) {
+			axios.delete(`http://localhost:8000/api/todo/${e.target.name}`,)
+			setTodolist((data) => {
+				return data.filter((todo) => todo._id !== e.target.name)
+			})
+		}
 	}
 
 	function handleUpdate() {
-		console.log("update:", update);
 		setUpdate(!update);
 	}
 
 	function handleCheck(e) {
 		e.target.checked ? setCompleted(true) : setCompleted(false)
+		console.log("Task completed set to ", completed)
 	}
 
 	return (
@@ -67,7 +75,7 @@ function App() {
 								<div className='flex justify-between items-baseline'>
 									<input value={completed} type="checkbox" className='outline-none' onChange={handleCheck} />
 									<div className='text-left'>
-										<h3 className={`h-12  ${completed ? 'line-through' : 'none'}`}>{data.title}</h3>
+										<h3 className={`h-12  ${completed ? 'line-through' : ''}`}>{data.title}</h3>
 										<code className='p-1 pr-2 text-xs rounded-md bg-blue-300 shadow-lg'><FontAwesomeIcon icon={faTag} className='px-2' />{data.category}</code>
 									</div>
 									<div className='flex flex-col w-20 p-2 m-2 gap-2'>
@@ -84,7 +92,7 @@ function App() {
 					</ul>
 				</section>
 				{open ? (
-					<div className="shadow-lg">
+					<div>
 						<p onClick={handleClose} className="close">
 							&times;
 						</p>
@@ -96,8 +104,39 @@ function App() {
 						/>
 					</div>
 				) : ("")}
+				{openModal ? (<Modal isOpen={openModal} style={{
+					overlay: {
+						position: 'fixed',
+						top: 30,
+						left: 50,
+						right: 50,
+						bottom: 30,
+						borderRadius: '6px'
+					},
+					content: {
+						position: 'absolute',
+						top: '230px',
+						left: '300px',
+						right: '300px',
+						bottom: '230px',
+						filter: 'drop-shadow(0 20px 13px rgb(0 0 0 / 0.03)) drop-shadow(0 8px 5px rgb(0 0 0 / 0.08))',
+						background: '#bfdbfe',
+						overflow: 'auto',
+						WebkitOverflowScrolling: 'touch',
+						borderRadius: '6px',
+						border: 'none',
+						outline: 'none',
+						padding: '20px'
+					}
+				}} >
+					<div>Are you sure you want to delete this task?</div>
+					<div className="flex justify-around items-center pt-6">
+						<button className='py-4 rounded-md bg-gray-100 px-20 shadow-xl' value="false" onClick={(e) => handleClick(e)} >No</button>
+						<button className='py-4 rounded-md bg-blue-400 px-20 shadow-xl' value="true" onClick={(e) => handleClick(e)} >Yes</button>
+					</div>
+				</Modal>) : ("")}
 			</div>
-			<div className="shadow-xl w-full bg-pink-700 rounded">
+			<div className="shadow-xl w-full bg-pink-300 rounded">
 				<CreateTodo handleUpdate={handleUpdate} />
 			</div>
 		</div>
